+++ internal/util/slices.go
@@ -5,3 +5,12 @@ func ConcatSlice[T any](first []T, second []T) []T {
 	n := len(first)
 	return append(first[:n:n], second...)
 }
+
+func Contains[T string | int](s []T, e T) bool {
+	for _, a := range s {
+		if a == e {
+			return true
+		}
+	}
+	return false
+}