+++ resources/profiles/spec.go
@@ -1,11 +1,12 @@
-package mmd
+package profiles
 
 // Spec is the configuration for MMD term set source
 type Spec struct {
+	// Whether to enable user profiles sync
+	Enabled bool `json:"enabled"`
 	// Optional, an alias for the table name
-	// Don't map different term sets to the same table - such scenario is not supported
 	Alias string `json:"alias"`
 }
 
-// SetDefault sets default values for MMD spec
+// SetDefault sets default values for list spec
 // func (s *Spec) SetDefault() {}